+++ src/cache_dit/utils.py
@@ -7,7 +7,6 @@ import contextlib
 
 from cache_dit.logger import init_logger
 
-
 logger = init_logger(__name__)
 
 
@@ -84,15 +83,3 @@ def print_tensor(
             f"{name}, mean: {x.float().mean().item()}, "
             f"std: {x.float().std().item()}, shape: {x.shape}"
         )
-
-
-def normalize_quantize_type(quantize_type: str | None) -> str | None:
-    if quantize_type is None:
-        return None
-    mapping = {
-        "float8_wo": "float8_weight_only",
-        "int8_wo": "int8_weight_only",
-        "int4_wo": "int4_weight_only",
-        "bnb_4bit": "bitsandbytes_4bit",
-    }
-    return mapping.get(quantize_type, quantize_type)
